RSA: Replace `RsaSubjectPublicKey::{modulus,exponent}` with `rsa::Public::{Modulus,Exponent}::be_bytes()`.
This is a step towards removing the heap-allocated and usually-unnecessary
`public_key: RsaSubjectPublicKey` field. The new API allows the caller to
better control how it stores/allocates the component values. This also removes
a couple of infallible `unwrap()`s.

This is a step towards removing `io::Positive` from the public API.

This is a breaking API change.