RSA: Split `rsa::KeyPair::sign` encoding and arithmetic.
Factor out the arithmetic so that it can be reused by the upcoming OAEP
implementation.

Remove the reference to the "Manger's Attack revisited" paper since it
is already cited in `fill_be_bytes` and because, at the time
`fill_be_bytes` is called by `sign()`, that attack is no longer
relevant since we're operating on the final signature value itself.