Remove leading blank lines in convert_comments.go
The OpenSSL style writes multiline block comments with a blank line at
the top and bottom, like so:

  /*
   * Some multi-line
   * comment
   */

The script already removed the trailing blank line, but not the leading
one. When we go to run this script in crypto/asn1, etc., we'll come
across those comments.

Change-Id: I189aec87a08607008779f883a97f2c53d24ee2da
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/52730
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: Bob Beck <bbe@google.com>