clang-format remaining directories.
Previously, we did not clang-format a few directories because we had
left them largely untouched. clang-format them now so we're finally more
uniform.

This CL is the result of the following commands:

    for d in asn1 x509 x509v3 pem; do
      clang-format -i crypto/$d/*.h
      clang-format -i crypto/$d/*.c
    done

(Written in this funny way because crypto/pem/*.h doesn't match
anything.)

Change-Id: I7f4ca9b3a9c8f07d6556e00e9e84b3c0880ee12e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/53085
Commit-Queue: Bob Beck <bbe@google.com>
Reviewed-by: Bob Beck <bbe@google.com>