CI: Fix Android builds in GitHub Actions.
When `$ANDROID_NDK_HOME` is set, `mk/cargo.sh` will use it to locate the
NDK tools. Thus, we need to ensure that the hack is applied to
`$ANDROID_NDK_HOME` when that is set.

I am not sure why I added this line; perhaps it was unintentially left in
when the script was being debugged.