Don't include <stdatomic.h> in C++
OPENSSL_C11_ATOMIC is both computed in crypto/internal.h and also
defined externally. This is a remnant of C11 atomics were an opt-in
feature.

If defined externally, this means OPENSSL_C11_ATOMIC might be defined
when built as C++. That, in turn, causes <stdatomic.h> to be included in
C++ mode. At least one of our users toolchains has a <stdatomic.h> that
is incompatible with C++.  We don't get anything out of including it, so
just gate the include on !defined(__cplusplus) for now.

Things to look into as follow-up:

- Fix build files to stop defining OPENSSL_C11_ATOMIC. Prior to
  https://boringssl-review.googlesource.com/c/boringssl/+/59847, it was
  still serving a purpose: in server builds, if autodetection fails, we
  would rather fail to build than accidentally fallback to locks.
  There is no lock fallback anymore.

- Fix that toolchain so their <stdatomic.h> is C++-compatible. It's
  certainly not C++23-conformant. I suspect it's also not
  C++11-conformant, but I'm not positive.

Change-Id: I13bcd8380efeb87b9f9cc439fe24a743e48aec60
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/59985
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Bob Beck <bbe@google.com>