Use sources.cmake for pki and pki_test
This achieves much of the same goals as
https://boringssl-review.googlesource.com/c/boringssl/+/61245, but
reuses the existing source.cmake parser in generate_build_files.py.

There are two practical differences here:

First, CMake knows that, if it sees include(sources.cmake), it should
add sources.cmake as a dependency of build.ninja, and re-run cmake if
source.cmake changes. It does not know this for file(STRINGS). This can
be fixed with CMAKE_CONFIGURE_DEPENDS, but since we already have this,
may as well use it.

Second, sources.cmake wants paths rooted at the top-level BoringSSL
source directory, which means we want to define the targets at the top
level, rather than in subdirectories. While it will make the top-level
CMakeLists.txt larger, I think this is a good move:

- ./build/crypto_test --gtest_filter=... is just less annoying to type
  than ./build/crypto/crypto_test --gtest_filter=...

- It will (eventually) mean libcrypto.a lives at build/libcrypto.a
  instead of build/crypto/libcrypto.a. This means we can pass a single
  -L parameter when building things against BoringSSL, without relying
  on the install target.

- It aligns with the generated CMake build. I would like, long-term, to
  stop maintaining multiple CMake builds, and the generated one seems to
  have a better convention. And one that will be more disruptive to
  others if we change it.

- Top-level-rooted paths are more convenient for generate_build_files.py
  to work with, because that's what we want to output.

As I get further along in 542, I expect I'll move this once again,
possibly to some JSON file, because I'll need my new pregenerate tool to
parse it (and we'll no longer be constrained by what's easy to consume
in CMake), but use this for now.

Bug: 542
Change-Id: I1e8b894057264da5d7624a1fbb92f9e1198ea38e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/61266
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: Bob Beck <bbe@google.com>
Auto-Submit: David Benjamin <davidben@google.com>