Make the parse_certificate_fuzzer from libpki work inside boring.
Seed the corpus from cert_corpus. As part of that, check in the result of minimizing all the corpora.

Note this is just making one of the fuzzers build, I'll adapt
the others and follow on by updating the IMPORT process to do it
in a follow on cl.

Bug: chromium:1322914

Change-Id: Iea1b89f8fee938fa99c0a4d8134bcd0e7023d149
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/61765
Commit-Queue: Bob Beck <bbe@google.com>
Reviewed-by: David Benjamin <davidben@google.com>