Name the hash functions in the Kyber implementation
When I was experimenting with different hash functions, it was useful to
split them out, and I think this is clearer anyway. E.g. there are two
uses of SHAKE-256 in the round3 submission, but the spec gives them
different names.

XOF doesn't get its own abstraction because we need to stream data out
of it.

ML-KEM similarly names things, though it will replace the KDF with a
hash J that only gets applied to the rejection value.

Change-Id: I67f1ac0216e05f61557023229572ff5fb36ed529
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/64407
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>