bigint: Improve temporary array alignment.
Replace pointer-to-usize `as` conversions with use of `align_offset`.
This avoids any dependency on the way pointers are encoded.

Previously when the buffer was already aligned the way we needed it,
we'd treat the buffer as being misaligned by `ALIGNMENT` bytes. This
wasn't problematic but it wasn't optimal. By using `align_offset`,
we'll recognize it as being misaligned by 0 bytes, i.e. already
aligned.

Previously we were allocating `ALIGNMENT` extra Limbs to have slack
for alignment where we could have allocated `ALIGNMMENT / LIMB_BYTES`
extra limbs. Now, since we make the minimum alignment adjustment
necessary, we can allocate 1 less limb than that.

Be explicit that the type of the buffer elements is `Limb`. Use
`core::mem::{align_of}::size_of::<Limb>()` instead of `LIMB_BYTES`.