Remove remnants of the old Android CMake toolchain
The PATH-related workarounds for the old third-party Android toolchain
don't seem to apply to the official NDK one, so just remove them. The
official one does define the ANDROID variable, but only in an
"android-legacy.toolchain.cmake". I think the standard way to do this is
CMAKE_SYSTEM_NAME so switch to that.

Change-Id: I3f3fa69d482893c333ae06f54ac39434c3c6a56c
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/65670
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: David Benjamin <davidben@google.com>