Use std::copy instead of OPENSSL_memcpy for the internal bssl::Array::CopyFrom
This (internal) abstraction was originally made for trivial types, but
if we ever got a complex type, we should use C++ copies, not C copies,
to preserve all the rules of that type.

A good STL will specialize std::copy to memmove/memcpy when possible, so
this should not appreciably change the generated code.

Change-Id: I76af334ef667e545dbbbe87315ce5b30a327358c
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/66427
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Bob Beck <bbe@google.com>