Use only predefined environment variables to compute name prefixes.
Instead of having build.rs set a custom environment variable, just
compute the value directly from environment variables that are
guaranteed to be set. This way, non-Cargo build systems can avoid
duplicating this custom logic.

Move the consistency check of `links` and the package version so that
the check is also done during pregeneration.

Use `read_env_var()` for the relevant environment variables so that
build.rs emits `cargo:rerun-if-env-changed`, as it was noticed that
Cargo otherwise doesn't seem to rerun the build script when the
Cargo.toml values are changed.