AES-GCM: Clarify safety of ARMv7 BSAES.
Although we never called `bsaes_ctr32_encrypt_blocks` with blocks == 0,
this wasn't obvious. Make it obvious. Document other safety requirements
specific to BSAES.

While doing so, change `ctr32_encrypt_blocks()` to be an unsafe function;
it is unsafe because it's safety relies on the caller ensuring that `f`
meets the contract that is now documented here.