AES-GCM internals: Use consistent definitions for `Counter`.
Use the same type definition for counters as BoringSSL and
`aes_nohw_ctr32_encrypt_blocks`. This should be a no-op for
correctness as the representation is effectively the same,
except now with looser (u8 vs u32) alignment. This is much
clearer though, and removing the type punning paves the way
for aes_nohw.c to be replaced with Rust code.