AES-GCM AArch64: Clarify integrated AES-GCM control flow.
When we use `aes_gcm_enc_kernel` it will process all the whole blocks,
leaving only the last partial block, if any. Clarify this in the
sealing implementation. With this change, the compiler understands
what's happening well enough to jump directly to processing the last
partial block after calling `aes_gcm_enc_kernel`.

Ideally we'd do the same for the opening implementation, but we can't
right now due to the support for overlapping input/output. We should
revisit this later when we have a better abstraction for that.