Don't add extra 'informational' errors in the delegate
So when IsPublicKeyAcceptable returns false, this unambigiously
adds the UnacceptablePublicKey error to the certificate.

However the delegate currently adds an Additional error to the
certificate, to add some information about what was there.

This causes problems with tests in chrome that care about
seeing the unacceptable public key as such, without the
certificate being tagged as just having multiple errors and
therefore generically invalid.

Instead of adding an extra error for information, add
this as a warning, then there's no problem.

Change-Id: Id24e913df17f6c8bbc63c19b31df3f80e59f638b
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/67888
Reviewed-by: David Benjamin <davidben@google.com>
Commit-Queue: Bob Beck <bbe@google.com>