[DEPS] Migrate from Chromium git to CIPD
The primary reason for this change is to resolve
https://crbug.com/boringssl/653. It has the side benefit of having much
more performant checkout times ( 1.5 mins to ~10 seconds locally).

Note that this change should be transparent to BoringSSL, but there
are some minor differences in the packages being pulled. Firstly,
we will not have Chromium-local changes. See the changelog here:

https://chromium.googlesource.com/android_ndk/+log

Additionally, this change goes from the base r23 to r23c.

Bug: crbug.com/boringssl/653
Test: Local checkout, then relying on infra
Change-Id: Id394eb254d00884c0e9aeb74c2103908c23fb8c3
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/68187
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: David Benjamin <davidben@google.com>