AES: Use `NonZero_size_t` for ctr32_encrypt functions.
It turns out BSAES isn't the only one that required a non-empty in/out;
at least the x86-64 implementation of `aes_hw_ctr32_encrypt_blocks`
requires a non-empty in/out as well. Just assume all of them require it.

As the code currently exists, this is a no-op because
`ctr32_encrypt_within` is never called with an empty in/out, but this
would prevent memory safety issues if/when that changes.