poly1305 x86_64: Remove poly1305_vec.c.
On x86_64, when we have SSE 4.1, which is pretty much always the case,
we use the integrated implementation. We also have the non-SSE
implementation to fall back on. This middle one is infrequently used
but creates difficulties with cross-compilation because of the use of
the <emmintrin.h> header file, which forces a dependency on stdlib.h
we wouldn't otherwise need.