cpu: Reduce scope of `unsafe` and clarify safety properties.
Completely separate the storage of flags on a per-architecture basis,
so that ARM/Aarch64 and x86/x86_64 code paths can begin to further
diverge in how they handle CPU feature flags.

Reduce the scope of each of the `static mut` declarations to the
minimum, and encapsulate the access to each in a minimal, safe
internal API.

Within the implementations of those new internal APIs, use
`addr_of!`/`addr_of_mut!` and `ptr::{read,write}` to access the
`static mut` fields.