aes_gcm: Prefer integrated implementations, avoid inlining fallbacks.
The `#[inline(always)]` on `open_strided` was just a mistake.

But, even correcting that, the compiler won't structure `open()` and
`seal()` in the expected way. With these changes, the integrated
implementations will get inlined (as far as possible) and the
compiler will be biased towards assuming they will be used.