Compute should_ack_sni in one place
Right now it's split over the SNI extractor and the code that calls the
SNI callback. But the only thing that reads it is the code to send the
ServerHello extension, so we don't need to half-compute it earlier.

Change-Id: Ia0fdd502ab21bec2ccdde99185951601a4650efe
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/69747
Reviewed-by: Bob Beck <bbe@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: Bob Beck <bbe@google.com>