Fix one of two issues breaking the Bazel build
Adding a C++ file into libcrypto triggered some more Bazel workarounds
in our build. For whatever reason, that's now being passed a "frozen"
list, so we cannot mutate the values.

This isn't sufficient to fix it, as we're also running into issues with
aes/aes.c being an internal "header". To fix that, I think we have to
bite the bullet and finally rename the FIPS fragments to .h, reflecting
the fact that they are intended to be #included.

Also the new Bazel seems to have a different MODULE.bazel.lock format,
so pick up that update.

Bug: 362664827
Change-Id: Ibf80fa389e0e1fd2b6f1ab110f0dea78a2a1a89b
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/70627
Commit-Queue: Adam Langley <agl@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>