Add an optional Bazel package to CI
Also condition all the tools the Bazel builders won't need. (Perl and Go
aren't build dependencies anymore.) Given we've broken this build a
couple times, it's clear we need to just put it on CI.

This will be paired with a new recipe in the recipes repo.

Bug: 42290610
Change-Id: I90a4b376849dee9dde3dfcd12c8df60db1b0120c
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/70747
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Bob Beck <bbe@google.com>