Fix delocate with the aarch64 NO_ASM build
This compiler sometimes emits code like this:

	adrp x10, .Llocal_data2+16
	ldr q0, [x10, :lo12:.Llocal_data2+16]

We transform it into:

	adr x10, .Llocal_data2+16
	ldr	q0, [x10]

Note this makes some assumptions on the compiler, which I've documented
in comments. We already have a similar assumption on ADRP + ADD pairs,
but it is a little more likely for the compiler to do this in the ADRP +
LDR case.

Hopefully we can get the delocate replacement working soon and make all
this moot.

Change-Id: Icf4ed701142a52edf38d285c0bc5d52c17032d4f
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/71267
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Adam Langley <agl@google.com>