Check for HelloRetryRequest in the ServerHello state
Previously it was a separate state because we stored the
received_hello_verify_request bit implicitly in the state. But now that
we've burned a bit for that, let's make that be one state.

This doesn't make a huge difference but avoids do_enter_early_data
implicitly assuming that DTLS 1.3 doesn't have 0-RTT. (0-RTT DTLS 1.3 is
not a priority. Rather I was getting caught up in the weird
SetVersionIfNullCipher silliness, which is linked with 0-RTT.)

Bug: 42290594
Change-Id: I28d5ea32657731bfbaff9c2bb9a02bfc33d8e41b
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/71527
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>
Auto-Submit: David Benjamin <davidben@google.com>