Add a bssl::InplaceVector to libssl
We're constantly hand-rolling these inlined bounded array structures.
Inspired by C++26's std::inplace_vector name, add something very basic.
This is still very barebones and probably can be improved. But it's
enough to be able to use spans more often.

Perhaps in a decade we'll be able to just use the real one. To keep our
size optimization, InplaceVector picks the minimal integer needed to
represent the size but, unlike doing it by hand, this is entirely hidden
from the API.

I switched a pile of fields over to this to confirm it mostly works,
though there are still a ton more that are array/length pairs. In
particular, we never came up with a good calling convention for all the
intermediate hash-sized values in the TLS 1.3 key schedule. I've left
that alone for now.

The immediate motivation here is that a lot of the transport-specific
epoch structs will need to store the original traffic secret, and
holding on to it everywhere without an abstraction is tedious.

Bug: 371998381
Change-Id: I17a987299dc9278b59d2b773e1f3fead69dff162
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/71748
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Reviewed-by: Bob Beck <bbe@google.com>