Remove redudant version information from SSLAEADContext
After https://boringssl-review.googlesource.com/c/boringssl/+/71533, it
is never the case that we have two version's worth of the record layer
alive at once. This means we stop needing to store a version number in
the epoch state, or worry about which version we check if in theory the
connection and each epoch are in different states. (I'm not sure how
DTLS 1.3 ACKs will possibly work in such a model.)

Simply this by taking versions out of the epochs and instead having
everything check the connection's version. For that to work, we need to
relax the invariants on the version slightly. When we enter 0-RTT, we
set the version to the 0-RTT version. The final version will then either
match or trigger a 0-RTT reject before changing the version. Having the
version change is unfortunate but it reflects what's going on.

As a result, some checks added for DTLS 1.3 will now behave correctly
with DTLS 1.3 0-RTT. This also simplifies things slightly along the way
to cleaning up the epoch state to implement DTLS 1.3 ACKs.

There are two checks that arguably did not want to pay attention to the
0-RTT version, but the difference seems unimportant. But, for
completeness:

1. If we offer 0-RTT, the logic to ignore dummy ChangeCipherSpec records
   deep in the record layer will kick in before the ServerHello, instead
   of after. This is arguably supported by RFC 8446, which says any time
   after the first ClientHello suffices:

   > An implementation may receive an unencrypted record of type
   > change_cipher_spec consisting of the single byte value 0x01 at any
   > time after the first ClientHello message has been sent or received
   > and before the peer's Finished message has been received and MUST
   > simply drop it without further processing.

   Though that guidance is pretty bizarre because we haven't learned the
   version yet before ServerHello. Anyway, I don't think this has any
   real practical impact and isn't worth fussing over.

2. If we offer 0-RTT, the logic to treat a "warning" decode_error alert
   as fatal will kick in before the ServerHello, instead of after. If
   the server deployed 0-RTT, then rolled back to TLS 1.2 (against the
   guidance in Appendix D.3 but possible), AND if it sent a warning
   alert before the TLS 1.2 ServerHello, we would fail the connection
   with a different error and not trigger the application's retry.

   This also does not seem worth fussing over.

Bug: 42290594
Change-Id: I8d97e219888c75274c3c640b2da5c769c6b9ad36
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/71534
Commit-Queue: Bob Beck <bbe@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Bob Beck <bbe@google.com>